requirements.txt
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
requests==2.26.0
py~=1.11.0
colorama~=0.4.4
pip~=22.0.4
attrs~=19.3.0
toml~=0.10.2
tornado~=6.2
Jinja2~=3.1.2
pytz~=2019.3
click~=7.1.2
setuptools~=58.1.0
pyflakes~=2.3.1
wcwidth~=0.1.8
pluggy~=0.13.1
iniconfig~=1.1.1
atomicwrites~=1.4.0
certifi~=2019.9.11
tzlocal~=4.2
six~=1.14.0
idna~=2.8
urllib3~=1.25.6
chardet~=3.0.4
telegram~=0.0.1
APScheduler~=3.6.3
cachetools~=4.2.2
pyparsing~=2.4.6
snowballstemmer~=2.2.0
SQLAlchemy~=1.4.37
legacy~=0.1.6
pydocstyle~=6.1.1
python-dateutil~=2.8.2